fixed low precision U map in 77259af059057f2892efd5a494d490cd46d0e8ec

+++ include/libint2/boys.h
@@ -786,7 +786,7 @@ namespace libint2 {
 
 
   namespace detail {
-  constexpr float pow10(long exp) {
+  constexpr double pow10(long exp) {
     return (exp == 0) ? 1. : ((exp > 0) ? 10. * pow10(exp-1) : 0.1 * pow10(exp+1));
   }
   }
